\name{util.390816171110-package}
\alias{util.390816171110-package}
\alias{util.390816171110}
\docType{package}
\title{
\packageTitle{util.390816171110}
}
\description{
\packageDescription{util.390816171110}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{util.390816171110}
\packageIndices{util.390816171110}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{util.390816171110}

Maintainer: \packageMaintainer{util.390816171110}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
